removes dead code and moves the new processMiddleware back to the file where the old one was.
