+++ src/types/middleware.ts
@@ -15,21 +15,10 @@ export type AnyMiddlewareArgs =
 export interface Context extends StringIndexed {
 }
 
-// export type ProcessMiddlewareContext<Args> = Args & {
-//   next?: NextMiddleware, context: Context,
-//   logger: Logger,
-//   client: WebClient,
-// };
-
-// export type MiddlewareContext<Args> = ProcessMiddlewareContext<Args> & {
-//   next: NextMiddleware,
-// };
-
 // NOTE: Args should extend AnyMiddlewareArgs, but because of contravariance for function types, including that as a
 // constraint would mess up the interface of App#event(), App#message(), etc.
 export interface Middleware<Args> {
   // TODO: is there something nice we can do to get context's property types to flow from one middleware to the next?
-  // (ctx: MiddlewareContext<Args>): Promise<unknown>;
   (args: Args & {
     next?: NextMiddleware,
     context: Context,
